test: improve splitting of unit/integration tests
