+++ .github/CONTRIBUTING.md
@@ -2,10 +2,8 @@
 
 ## Versioning
 
-Follows [Semantic Versioning][semver-2]
+Follows [Semantic Versioning][semver]
 
-[github-flow]: https://docs.github.com/en/get-started/using-github/github-flow
-[semver-2]: https://semver.org/
 
 ## Code changes
 
@@ -13,23 +11,22 @@ Follows [GitHub flow][github-flow]
 
 ### Before commit
 
-Functional tests:
-
 ```bash
 npm test
-npm run integration-test
-```
-
-Meta:
-
-```bash
 npm audit
 npm run lint
 ```
 
-all of the above should exit with `0`; no errors and no warnings.
+all should exit with exit-code:`0`, without *any* errors or warnings.
 
 ## CI Workflows
 
-- Functional unit-tests should be run in the `test` workflow.
-- Non-functional tests, i.e `npm audit` should be run in the `meta` workflow.
+- [Functional tests][func-req] should be run in the `test.yml` workflow.
+- [Non-functional tests][non-func-req], i.e `npm audit` should be run in the
+  `meta.yml` workflow.
+
+
+[github-flow]: https://docs.github.com/en/get-started/using-github/github-flow
+[semver]: https://semver.org/
+[func-req]: https://en.wikipedia.org/wiki/Functional_requirement
+[non-func-req]: https://en.wikipedia.org/wiki/Non-functional_requirement
