+++ .github/workflows/test.yml
@@ -22,6 +22,5 @@ jobs:
         cache: 'npm'
     - run: npm ci
     - run: npm test
-    - run: npm run integration-test
       env:
         NODE_ENV: test
