+++ test/index.spec.js
@@ -1,21 +1,8 @@
-// Instantiation
+// unit tests
+import '.unit/test/index.spec.js'
 
-import './new.spec.js'
-
-// lists
-
-import './lists/list/index.spec.js'
-import './lists/lazy-list/index.spec.js'
-import './lists/append-list/index.spec.js'
-
-// `map`
-
-import './map/flatten/index.spec.js'
-
-// repository
-
-import './repository/index.spec.js'
+// integration tests
+import '.unit/integration/index.spec.js'
 
 // test model
-
 import './model/test/index.spec.js'
