+++ test/unit/index.spec.js
@@ -1,0 +1,17 @@
+// Instantiation
+
+import './new.spec.js'
+
+// lists
+
+import './lists/list/index.spec.js'
+import './lists/lazy-list/index.spec.js'
+import './lists/append-list/index.spec.js'
+
+// `map`
+
+import './map/flatten/index.spec.js'
+
+// repository
+
+import './repository/index.spec.js'
