+++ test/unit/map/flatten/index.spec.js
